refactor: Add public function interface
Move all function logic into `internal/fn` and add a public
constructor in `pkg/fn` that allows to import and create a p&t function
in another module while keeping the function logic private.
