+++ internal/fn/connection.go
@@ -1,4 +1,4 @@
-package main
+package fn
 
 import (
 	"k8s.io/apimachinery/pkg/runtime"
