+++ internal/fn/doc.go
@@ -1,0 +1,2 @@
+// Package fn contains internal function implementation.
+package fn
