+++ internal/fn/validate.go
@@ -1,4 +1,4 @@
-package main
+package fn
 
 import (
 	"fmt"
