+++ main.go
@@ -5,6 +5,8 @@ import (
 	"github.com/alecthomas/kong"
 
 	"github.com/crossplane/function-sdk-go"
+
+	"github.com/crossplane-contrib/function-patch-and-transform/pkg/fn"
 )
 
 // CLI of this Function.
@@ -24,7 +26,10 @@ func (c *CLI) Run() error {
 		return err
 	}
 
-	return function.Serve(&Function{log: log},
+	return function.Serve(
+		fn.NewFunction(
+			fn.WithLogger(log),
+		),
 		function.Listen(c.Network, c.Address),
 		function.MTLSCertificates(c.TLSCertsDir),
 		function.Insecure(c.Insecure))
